+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## [2026.4.1]
+
+### Added
+
+- Designed abstract class for AnimeLibrary component
+- Added kernel method showsInCurrWatched
+- Added kernel method showsInWatchlist
+
 ## [2026.3.9]
 
 ### Added